add comments describing behaiour of code with TODO lines describing some potential bugs it may be you want the code to work that way, but best be clear.
renamed some variables for clarity and to stop clobbering buffer.
I have not renamed to highlight scope. as there is the same variable names
clobering each other in different scopes.
the variables you are using for this are constants so wont change and so doesnt
make any difference but, then you are using them as globals anyway so yeah.